+++ lib/src/tab_button.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:tabbed_view/src/icon_path.dart';
 import 'package:tabbed_view/src/tabbed_view_menu_builder.dart';
+import 'package:tabbed_view/src/theme/tabbed_view_theme_constants.dart';
 
 /// Configures a tab button.
 class TabButton {
@@ -13,7 +14,11 @@ class TabButton {
       this.onPressed,
       this.menuBuilder,
       this.toolTip,
-      this.padding}) {
+      this.padding,
+      double? iconSize})
+      : this.iconSize = iconSize == null
+            ? iconSize
+            : TabbedViewThemeConstants.normalize(iconSize) {
     if ((this.iconData != null && this.iconPath != null) ||
         (this.iconData == null && this.iconPath == null)) {
       throw ArgumentError('iconData or iconPath must be set');
@@ -29,4 +34,5 @@ class TabButton {
   final VoidCallback? onPressed;
   final TabbedViewMenuBuilder? menuBuilder;
   final String? toolTip;
+  final double? iconSize;
 }