+++ lib/src/theme/tabbed_view_theme_constants.dart
@@ -1,4 +1,11 @@
 class TabbedViewThemeConstants {
   static const double minimalIconSize = 8;
   static const double defaultIconSize = 10;
+
+  static double normalize(double buttonIconSize) {
+    if (buttonIconSize >= TabbedViewThemeConstants.minimalIconSize) {
+      return buttonIconSize;
+    }
+    return TabbedViewThemeConstants.minimalIconSize;
+  }
 }